fix: clean up helpers
