+++ apps/platform/src/components/AssociationsToolkit/utils/associations.js
@@ -1,0 +1,99 @@
+import { ENTITIES } from "./index";
+import { v1 } from "uuid";
+
+const getEmptyRow = id => ({
+  dataSources: {},
+  score: 0,
+  disease: { id },
+  target: { id },
+});
+
+/***********
+ * HELPERS *
+ ***********/
+export const diseaseAssociationsTargetSelector = data =>
+  data[ENTITIES.DISEASE].associatedTargets.rows;
+export const targetAssociationsDiseaseSelector = data =>
+  data[ENTITIES.TARGET].associatedDiseases.rows;
+
+export const diseasePrioritisationTargetsSelector = data =>
+  data[ENTITIES.TARGET].prioritisation.items;
+
+export const getPrioritisationData = data => {
+  const dataRows = diseasePrioritisationTargetsSelector(data);
+  const prioritisations = dataRows.reduce(
+    (acc, curr) => ((acc[curr.key] = parseFloat(curr.value)), acc),
+    {}
+  );
+  return { prioritisations };
+};
+
+export const getDataSourcesData = data => {
+  const sources = data.datasourceScores.reduce(
+    (acc, curr) => ((acc[curr.componentId] = curr.score), acc),
+    {}
+  );
+  return sources;
+};
+
+export const getDataRowMetadata = (parentEntity, row, fixedEntity) => {
+  let targetSymbol, diseaseName, id;
+  switch (fixedEntity) {
+    case ENTITIES.DISEASE:
+      id = row.target.id;
+      targetSymbol = row.target.approvedSymbol;
+      diseaseName = parentEntity.disease.name;
+      break;
+    case ENTITIES.TARGET:
+      id = row.disease.id;
+      targetSymbol = parentEntity.target.approvedSymbol;
+      diseaseName = row.disease.name;
+      break;
+    default:
+      return { targetSymbol, diseaseName };
+  }
+  return { targetSymbol, diseaseName, id };
+};
+
+export const getAllDataCount = (fixedEntity, data) => {
+  switch (fixedEntity) {
+    case ENTITIES.TARGET:
+      return data[ENTITIES.TARGET].associatedDiseases.count;
+    case ENTITIES.DISEASE:
+      return data[ENTITIES.DISEASE].associatedTargets.count;
+    default:
+      return 0;
+  }
+};
+
+export const getAssociationsData = (fixedEntity, data) => {
+  if (!data) return [];
+  const withPrioritisation = fixedEntity === ENTITIES.DISEASE;
+  const dataRows =
+    fixedEntity === ENTITIES.DISEASE
+      ? diseaseAssociationsTargetSelector(data)
+      : targetAssociationsDiseaseSelector(data);
+
+  return dataRows.map(row => {
+    const dataSources = getDataSourcesData(row);
+    const { targetSymbol, diseaseName, id } = getDataRowMetadata(data, row, fixedEntity);
+    return {
+      score: row.score,
+      id,
+      targetSymbol,
+      diseaseName,
+      dataSources,
+      ...(!withPrioritisation && { disease: row.disease }),
+      ...(withPrioritisation && { target: row.target }),
+      ...(withPrioritisation && getPrioritisationData(row)),
+    };
+  });
+};
+
+export const getInitialLoadingData = (rowCount = 10) => {
+  const arr = [];
+  for (let i = 0; i < rowCount; i++) {
+    arr.push(getEmptyRow(v1()));
+  }
+  return arr;
+};
